game/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiofiles==23.2.1